+++ erts/emulator/beam/erl_driver.h
@@ -576,7 +576,7 @@ EXTERN int driver_send_term(ErlDrvPort ix, ErlDrvTermData to,
 
 /* Async IO functions */
 EXTERN long driver_async(ErlDrvPort ix,
-			 unsigned long* key,
+			 unsigned int* key,
 			 void (*async_invoke)(void*), 
 			 void* async_data,
 			 void (*async_free)(void*));
