+++ erts/emulator/test/driver_SUITE_data/peek_non_existing_queue_drv.c
@@ -127,7 +127,7 @@ static int control(ErlDrvData drv_data,
 		   char **rbuf, int rlen)
 {
     PeekNonXQDrvData *dp = (PeekNonXQDrvData *) drv_data;
-    unsigned long key = 0;
+    unsigned int key = 0;
     char *res_str = "ok";
     ErlDrvSysInfo si;
     driver_system_info(&si, sizeof(ErlDrvSysInfo));
